% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apportion_scores_comp.R
\name{apportion_scores_comp}
\alias{apportion_scores_comp}
\title{Apportions outlier scores to composites.}
\usage{
apportion_scores_comp(obj)
}
\arguments{
\item{obj}{The output of \code{comp_tsout_ens}}
}
\value{
A list with the following components:
\item{\code{scores_ori}}{The apportioned outlier scores.}
\item{\code{scores_scaled}}{The apportioned outlier scores rescaled so that the sum of the composite outlier scores is equal to the total outlier score at that time point.}
}
\description{
Apportions outlier scores to composites.
}
