% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanish_morte-data.R
\docType{data}
\name{spanish_morte}
\alias{spanish_morte}
\title{A dataset containing mortality counts in Spain by state.}
\format{
A list of two items.
\describe{
 \item{1}{The compositional dataset of mortality proportions by state.}
 \item{2}{The relevant dates.}
}
}
\usage{
spanish_morte
}
\description{
This dataset contains mortality counts in Spain by state from the 18th of April 2018 until the 5th of May 2020.
}
\keyword{datasets}
