% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decomposed.R
\name{plot_decomposed_all}
\alias{plot_decomposed_all}
\title{Plot all decomposed time series from comp_tsout_ens or mv_tsout_ens output.}
\usage{
plot_decomposed_all(obj, X = NULL)
}
\arguments{
\item{obj}{The output from \code{comp_tsout_ens} or \code{mv_tsout_ens} functions.}

\item{X}{The data matrix used as input to \code{mv_tsout_ens} (not needed if \code{obj} is output from \code{comp_tsout_ens}).}
}
\value{
A ggplot showing the time series with facets by decomposition method.
}
\description{
Plot all decomposed time series from comp_tsout_ens or mv_tsout_ens output.
}
